Worked around branch bugs caused by memoryEvictCached() side effects.
StoreEntry::release() (during disk rebuild) and storeGetMemSpace() were
purging unlocked entries from the memory cache without abandon()ing
them. Purging an entry from a memory cache destroys mem_obj and, hence,
its Transients (if any). Such silent Transients loss violates the
public=hasTransients invariant and leads to bugs, especially if
StoreEntry was not cached in memory and, hence, did not broadcast its
eviction.

Also do not memoryEvictCached() in StoreEntry::release() because all
memoryEvictCached() calls are dangerous, and that call was not needed as
releaseRequest() yields to a (safe and broader) evictCached() call.

Dangerous purgeMem() is gone, and memoryEvictCached() is now private.

Also possibly fixed purging of incomplete on-disk entries that managed
to lose their mem_obj (in non-SMP caching environment):
Controller::handleIdleEntry() was not releasing such entries because it
relied on purgeMem() that did nothing without a mem_obj.