Limit 'selective collapsed forwarding' only to required contexts
The previous attempt (1ccd140) was reworked.  Passing boolean variable
for requesting collapsing analysis to low-level getPublic* methods
would work not well because it is not possible to construct proper
ACLFilledChecklist object without the context details. With this
approach, such objects are constructed directly at each context
and passed to StoreEntry::getPublic* methods.

Icp(and Urn) contexts now are analysed similarly to htcp ones(1ccd140
lacked that).