Remove redundant cstdarg includes
cstdarg is pulled into all src/ code through compat/stdvarargs.h in squid.h.

This intentionally leaves purge tool using it directly, which should not be a
problem and allows it to be built standalone.

Also, leaves the stdarg.h include in snmplib/snmplib_debug.c which needs to
be upgraded to C++ first.