Refactor squid.conf test design
This forces dev to manually maintain the list of squid.conf tests
in future, but allows the certificate tests to be omitted in
minimal builds where they are not relevant.

generate-test-certificates automake target is left non-conditional
to simplify the Makefile.am logic but does not produce an error
unless at least one of the certificate creation tools is provided
and an operation the script requires actually fails.