Identify collapsed transactions
Added a new COLLAPSED to the squid request status %Ss access log field.
This tag marks transactions that waited for the CF initiator transaction
in three possible cases:

1. Classic: client request gets collapsed on arrival, e.g.,
   TCP_COLLAPSED_HIT or TCP_COLLAPSED_MISS.
2. Collapsed revalidation: an internal revalidation request for
   a regular client request gets collapsed, e.g.,
   TCP_COLLAPSED_REFRESH_MODIFIED.
3. Mixed: an already collapsed client gets collapsed on revalidation,
   This case is marked like (2), e.g., TCP_COLLAPSED_REFRESH_MODIFIED.

These changes do not affect CF initiating code.

Also addressed a LogTags TODO, eliminating the copy assignment operator.

The given approach with a single COLLAPSED tag is simple but lacks
important information for some cases, for example:

* Obviously, (2) and (3) get identical tags and cannot be distinguished.
* A collapsed client which becomes a revalidation initiator cannot be
  distinguished from a not-collapsed but revalidation-collapsed client.

We can try to improve this in the future.