Fixed the XXX of 106198b commit message
The proposed fix uses the following assumption: if a caller creates
ACLChecklist with an ALE, it naturally expects that the ALE would be
used by some of ACLs of that list and that the ALE already has all
available context information (i.e., synced with the context).

With fix applied, any ACLChecklist with ALE synchronizes it at
'pre-check' stage without level-1 warnings triggering. Warnings are
triggered only if for some reason this 'pre-check' synchronization was
bypassed.