Fixed the XXX of 106198b commit message (second attempt).
The first try fixing it in 00fc3fa3 was unsuccessful because some new
problems were found in ACLFilledChecklist::syncAle() implementation that
approach did not care of. It is not possible to correctly initialize
some of ALE fields in syncAle() because ACLFilledChecklist lacks some
important data for it. For example, AccessLogEntry::url improperly
gets the current request URL (probably after redirection) instead of
log_uri ("ru" format code). So, silently initializing missing ALE fields
at pre-check stage with available, but possibly wrong values would be a
bad idea: the user even would not be warned about these problems.

This is another approach: fill ALE with the data ACLFilledChecklist is
not aware of (adapt_request and url) directly in the ACLFilledChecklist
caller context. There are lots of places where ACLFilledChecklist
initializes its ALE fields.  This patch does not fix them all, but only
those having the required data at hand. I hope the proposed fix would be
enough to hide the most of annoying warning messages.