Converted the StoreEntry::callCreated() method into a static function.
The awkward method avoided code duplication and granted access to
private StoreEntry members. However, it will have to be replaced when
StoreClient::created() is finally gone, and it did not need access to
private members.

A function is a better fit for the job, and we can still grant it
privileged access rights if needed (possibly converting it into a
private method at that point).