ext_edirectory_userip_acl: fix multiple strncat() buffer overreads
GCC 8 reports:
  error: strncat() specified bound 2 equals source length
 [-Werror=stringop-overflow=]

The right way to deal with IP address conversion is to work with the
binary representation and convert from that to the final output form.
Manipulating them as strings is very complex and prone to errors such
as these string buffer bounds issues.

Replace string handling in ConvertIP() with a conversion to binary
using the available addrinfo(), then print the binary as hex with
the new makeHexString() function.

Replace complex sequences of strncat() in SearchFilterLDAP() with
a simpler sequence of snprintf() which does input validation and
return an appropriate LDAP error on overflows.

Also, fix an incorrect bounds check in BindLDAP()