Finished moving StoreClient::collapsedStats into ALE
Removed StoreClient::collapsedStats and related in-branch leftovers.

While LogTags stores both collapsedStats and LogTags_ot tags, the two
pieces of information are quite different and should be updated at
different times (preferably as soon as the value of each piece is
determined or changed). The update() method no longer binds them.

BTW, the same can be set about the individual tags themselves (e.g., the
TCP tag should be set much sooner than HIT), but polishing that further
is outside this project scope.

htcpSyncAle() and icpSyncAle() should not require the caller to supply
correct LogTags because callers like fillChecklist() do not have access
to that information (it is not stored in the transaction object unlike
the other pieces of info that these functions copy to ALE).

BTW, htcpLogHtcp() and icpLogIcp() serve dual purpose: Each records
the final caller tags into ALE and logs the transaction (if needed).
Only the latter goal is reflected in function names.

I would have preferred to make fewer ICP changes, but I realized that
819be28 forgot ALE when queuing responses and was not sure my overall
approach could still deliver the tags to ALE::cache.code when that bug
is fixed, so I fixed the bug. Some of these changes are described below.

Added icpUdpData::ale to preserve master transaction info when messages
are queued. This should have been done in 819be28. Several icpUdpData
improvements were triggered by this change because ale is a (second!)
non-POD member and icpUdpData was mistreated as a POD. They include:

  * Removed icpUdpData::start as unused.
  * Removed icpUdpData::len as set but otherwise unused.
  * Removed icpUdpData::logcode as essentially duplicating msg->opcode.

Update ICP ALE, if any, as soon as the transaction tags become known
(instead of sometimes waiting for the ICP message to be logged). The
ICP message may be dropped and/or never be logged, but we should keep
ALE up to date because it is used in an increasingly many contexts. I
doubt this change fixes any "real" bugs today (because ICP code does
not use ALE much), but it allowed us to hide icpUdpData in icp_v2.cc.

Also found and marked an ICP memory leak. It is best to fix that in a
dedicated commit.