Fixed branch bug: collapsedStats were incremented too early
The hit recipient (e.g., a StoreClient::created() method) often has
several preconditions for using cache hits. If a precondition fails, the
transaction does not collapse on the received entry (it follows a cache
miss code path instead). Thus, we cannot increment collapsedStats just
because there was a cache hit with an appropriate cache entry. That
increment now happens in startCollapsingOn().

It is tempting to remove StoreClient::loggingTags() now because all
startCollapsingOn() callers can supply their collapsedStats counter
directly to that method as a parameter.

I did not remove loggingTags() because it is possible that we should
restructure created() and related code so that it is difficult to forget
adding important hit preconditions such as collapsed_forwarding_access
(two such forgotten places have been identified and marked with XXXs in
this commit). The restructured code will need loggingTags(). Such
restructuring is too big for this logging/reporting project.

Also renamed collapsingInitiator() and polished its documentation so
that it can be more comfortably reused in more contexts. The method does
not necessarily identify a StoreEntry object that _initiated_
collapsing. It identifies a StoreEntry object using which _requires_
collapsing (because the object has not seen response headers yet and,
hence, may require the caller to send a brand new request when those
headers do arrive and prohibit sharing).