Update cppunit test documentation
Update the unit test template to clarify what the best layout should
be for simplest dependencies, fewer automake dist actions and more
compact structure for the relevant automake variable contents.

Also, move the TESTS assignment to follow all check_PROGRAMS
alterations. This is not strictly necessary, but makes it clear that
*all* check_PROGRAMS contents are being run, not the subset that are
above it.

Also, update the Boilerplate example test to comply with the current
documentation and move it up to alphabetical order (by test binary
name) following the template documentation.