Identify collapsed transactions (#213)
Added a new CF to the squid request status %Ss access log field.
This tag marks transactions that waited for the CF initiator
transaction in three possible cases:

1. Classic: client request gets collapsed on arrival, e.g.,
    TCP_CF_HIT or TCP_CF_MISS.
2. Collapsed revalidation: an internal revalidation request for
    a regular client request gets collapsed, e.g.,
    TCP_CF_REFRESH_MODIFIED.
3. Mixed: an already collapsed client gets collapsed on revalidation,
    This case is marked like (2), e.g., TCP_CF_REFRESH_MODIFIED.

These changes do not affect CF initiating code.

The given approach with a single CF tag is simple but lacks
important information for some cases, for example:

* Obviously, (2) and (3) get identical tags and cannot be distinguished.
* A collapsed client which becomes a revalidation initiator cannot be
  distinguished from a not-collapsed but revalidation-collapsed client.

We can try to improve this in the future.


In order to track collapsed transactions, a new CollapsingHistory
class was introduced. Since more and more non-logging code relies
on ALE, this new data is keeped in ALE along with other historical
data. ClientHttpRequest uses its logType field instead of the LogTags
in ALE, so we also use logType for storing CollapsedHistory. Eventually,
ClientHttpRequest should eliminate logType in favor of direct ALE use.

Also: ICP code fixing/refactoring:
* htcpSyncAle() and icpSyncAle() should not require the caller to supply
  correct LogTags because callers like fillChecklist() do not have
  access to that information (it is not stored in the transaction object
  unlike the other pieces of info that these functions copy to ALE).
* Added icpUdpData::ale to preserve master transaction info when
  messages are queued. Several icpUdpData improvements were triggered by
  this change because ale is a (second!) non-POD member and icpUdpData
  was mistreated as a POD. They include:
  - Removed icpUdpData::start as unused.
  - Removed icpUdpData::len as set but otherwise unused.
  - Removed icpUdpData::logcode as essentially duplicating msg->opcode.
* Update ICP ALE, if any, as soon as the transaction tags become known
  (instead of sometimes waiting for the ICP message to be logged). The
  ICP message may be dropped and/or never be logged, but we should keep
  ALE up to date because it is used in an increasingly many contexts.

Also: found and marked an ICP memory leak. It is best to fix that in a
dedicated commit.

Also: supply URN code with ALE. Full-featured Client-based classes
already use ALE. We have not tested with URNs, but these changes may
improve logging of transactions that involve URN resolution.

Also: fixed problematic StoreEntry::collapsingInitiator(). In fact,
it could return true if the entry has transients but has nothing to
do with collapsing. Also, it incorrectly assumed that a collapsed
entry is always marked with ENTRY_FWD_HDR_WAIT. However, this is not
true because Controller::allowCollapsing() does not set this flag
for the entry. This change introduced another improvement:

* separate collapsed hits from regular hits with a a new StoreEntry
  flag.  By the definition, collapsed hits can be served only by entries
  having this new ENTRY_REQUIRES_COLLAPSING flag. We have to add the
  new flag because there is no suitable existing set of
  methods/conditions for this purpose.