The %>handshake logformat code
Initial client bytes received by Squid on a newly accepted TCP connection
or inside a just established CONNECT tunnel.

For HTTP clients, the handshake is the request line.
For TLS clients, the handshake consists of all TLS records up to and
including the TLS record that contains the last byte of the first
ClientHello message.
For clients using an unsupported protocol, this field contains the bytes
received by Squid at the time of the handshake parsing failure.

Current support is limited to these contexts:
  - http_port connections, but only when the on_unsupported_protocol
    directive is in use.
  - https_port connections (and CONNECT tunnels) that are subject to
    the ssl_bump peek or stare action.

To protect binary handshake data, this field is always base64-encoded
(RFC 4648 Section 4).

This is a Measurement Factory project