Ignore Content-Length in 1xx or 204 responses
9dc4a88e already fixed Squid sending code. Since storing improper
Content-Length may confuse different Squid helpers/services/admins, it
is better to ignore it at the beginning, while parsing such responses.

Basically, ignore Content-Length in the receiving code should be enough.
However, we decided to leave removing irrelevant Content-Length in the
sending code for safety reason, because this check is fast and 1xx/204
responses are rare.