Use SBuf to collate the FD description for pconn
The persistent connection poool keys used in FD
description include domain names, which can at times far
exceed the FD_DESC_SZ space.

GCC-8 does not like truncation and complains with a warning
when -O3 optimizations are enabled. When combined with
-Wall -Werror as Squid does this halts compiling.

The alteration to dynamic descriptions for FD now allows
longer descriptions to be stored. SBuf is our long-term
buffer intended to hold this type of string so use it
already for this allocation and copy.