Undo branch commit 91ab5f8
... because it broke the basic "stare all/bump all" use case.

I do not know which 91ab5f8 changes are responsible for that breakage.
I know that FwdState::connectStart() change immediately triggers the
BUG message inside the modified if statement, and that undoing that
change alone is not sufficient to restore stare all/bump all" support.

It is possible that most of the other 91ab5f8 changes are correct, but
they need to be isolated and tested to weed out the bad apple. Also,
sslPeek/sslBumped changes/comments seem to mismatch sslPeek/sslBumped
flag descriptions in RequestFlags.h. If the changes/comments are
correct, then those descriptions need to be fixed.

XXX: Without 91ab5f8, the following SslBump configurations fail basic
tests: bump at step1, client-first, and server-first.