Bumped connections on step1 aborts with "BUG: Lost previously ..." message
The bumped connections on step1 are equivalent to client-first bumping mode
however are not handled with the same way inside FwdState::connectStart
where checks if a request marked as bumped received inside a bumped connection.

In these cases the ConnStateData::bumpMode is bumpBump but the is not any
ConnStateData::sslServerBump object.

Alternate solutions:
  - convert bumpBump action on step1 to bumpClientFirst action
  - or try build ConnStateData::sslServerBump object in all cases of bumping
    and check the ConnStateData::sslServerBump.act.step1 for bumpClientFirst
    or bumpBump modes.