Fixed stuck a4.a6.up6.happy.test case (and other improvements)
The a4.a6.up6.happy.test case got stuck because
HappyConnOpener::checkForNewConnection() incorrectly assumed that if we
started waiting for the activeSpareCall, then we should continue to wait
for it. That assumption is wrong when no spare paths are going to be
available. In that specific test case, we should not even started
waiting, but a different test case can be constructed where we cannot
tell that no wait will be needed when we start waiting.

Also, checkForNewConnection() quit if n_tries reached maxTries, but that
felt wrong because we may be already connecting (i.e., already using one
of those n_tries).

The checkForNewConnection() method is (or should be) used as a general
checkpoint (after any external changes) and should not make too many
assumptions about its caller. I started rewriting it accordingly.

The rest are comments for future work and minor polishing.