Reduce "openssl pollution" in high level code
Header compat/openssl.h is no longer included by compat/compat.h.
It is included only for headers that use openssl.

Openssl headers were removed from .cc files and placed in the
corresponding .h file, if actually needed.
If a header is not needed for function prototypes, but is used in
their definitions, it is included by the .cc file.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>