Introduced "recovery" stage
Happy Eyeballs RFC focuses on what to do with the first IPv4 and the
first IPv6 address. It is mostly silent regarding what happens when
there are multiple same-family addresses and the first address fails.
Yet, that post-failure stage is actually tricky to get right. This
commit is a (failed) attempt to do so.

The old code revolved around the idea that prime failure makes the spare
attempt the primary one, but there were several implementation bugs and
the primary/spare code was difficult to follow while thinking about
various combination of failures. Tests were failing.

To simplify, I introduced "recovery" stage dedicated to post-failure
handling of remaining IP addresses. The code is now easier to grok
because pre-spare, prime+spare, and recovery stages are better isolated.
Recovery code just round-robins through all the remaining IP address
families, combining fair family treatment with a simple implementation.

XXX: This is not the final solution. Tests (and contemplation) show that
family round robin may sadden user eyeballs once IPv6 is enabled, which
is the opposite of what Happy Eyeballs is about! Most users will not
notice the initial spare waiting gap of, say 250ms, but they will notice
increased TCP delays if IPv6 connections slowly timeout in combination
with a fast failure of the first IPv4 address(es):

* Without IPv6, fast IPv4 failure is largely unnoticeable to eyeballs.
  If the first IPv4 address quickly rejects the connection but the
  second IPv4 address works fine, then the user will think that
  everything is working OK.

* Once IPv6 is enabled, slowly failing IPv6 connection attempts may
  significantly delay the previously fast IPv4 recovery if the next IPv4
  attempt has to wait (due to round robin across families) for the
  previous IPv6 attempt to complete (i.e. timeout).