Fixed "we are waiting for an open connection" state detection
The old branch code was using connOpener.valid() which is a wrong test
because the opening job may be gone while we are still waiting for its
result being delivered to us. The test was OK (albeit optional) as far
as sending a message to the job is concerned, but not OK for testing
FwdState state.

Branch commit 651260e introduced a "(connOpener)" state test that did
not compile because CbcPointer does not have a default conversion to
boolean (not sure why Must(connOpener) works).

We are now using a dedicated method to wrap/hide connOpener.set() tests.