Refactor unit test layout to allow easier top-down analysis
... of what exists and what is still missing.

Definitions of testFoo binaries are grouped by the src/* subdir
their code resides in and ordered per the SUBDIRS listing so
that simpler code gets tested before complex code that uses it.

This way we have a definitive structure to the groupings (SUBDIRS)
to instruct new developers while also ensuring that complex code
has its dependencies tested for problems before it gets tested.

So we can hopefully avoid debugging complex cases just to find
that the simpler code it depended on was actually broken.

* Use only "+= foo" to add check_PROGRAMS entries for tests.

Having the list partially separated and partially in one location,
but with that location separated from the test definition has led
to difficulty at times determining whether a test is actually run.

We should now only use the += construct to build the list and
place that with the other values definign the unit test binary.

NP:
 This commit is purely a shuffling of lines within Makefile.am
 no lists other than check_PROGRAMS were changed and that only
 changes in what order tests are built and run.