Fix commit 70b53312
The commit 70b53312 did not set the HttpStateData::_peer in the case of
not "proxying", consider this variable useless in the case we are speaking
to an origin server even if we are using a tunnel through a cache peer.
In most cases this is looks that it is more-or-less correct but
looks confusing.

This patch:
- Remove the RequestFlags::proxying flag. The httpstateflags has a similar
  flag which should used instead.
- Inside HttpStateData class set the proxying flag when (a) a cache peer is
  used, (b) it is not an HttpRequest tunneled through the peer, and the
  peer is not an origin server.
  The proxying flag is set when the request destined to the peer, and not
  to the origin server (even if a peer used to tunnel it).
- Try to correctly set other flags and counters (eg keepalive related)
  based on flags.proxying.