Re-break the ASCII CTL character / OBSTEXT detection
The old stringHasCtl() only matched true for the UTF-8
characters 0x80-09f.

The encoder will correctly encode the full OBSTEXT range
*if* called, but the URL cleaner would not do that call
when only UTF-8 characters in the upper non-Latin-1
range(s) were used in a URL.

Fixing this is left out of scope for now. Added TODO
comments to the code highlighting the issue.