!Comm::MonitorsRead assertion in HttpStateData::maybeReadVirginBody()
Http requests on bumped (and pinned) connections using the server-first,
peek or stare bump mode must not retried or reforwarded.
This is because in all of these cases the client-to-squid TLS connection
established after evaluating and using information from the squid-to-server
TLS connection.
Squid should forward an error to the client instead.

Also the peer select code should return only one PINNED path.

This is a Measurement Factory project.

This is the t4 patch for this issue.