Removed reserveSlotForWriting() parameters as promised in last branch commit
Do not abuse StoreMap::writeableSlice() for the free slot management:
That slot is not associated with any entry chain (yet) and that method()
is declared to apply to in-chain slices only.

Long-term, we should move freeSlots into StoreMap, along with
the primary reserveSlotForWriting() logic. All StoreMap users that
use freeSlots follow the same overall logic AFAICT.