Upgrading to a new protocol does not work when REQMOD adaptation is used
This is because the Upgrade header removed from HttpRequest object, because
it is a hop-by-hop header. The Upgrade header removed before sent to an
adaptation service, and the adaptation service normally will not include
a such header.
We need to access virgin HttpRequest header in HttpStateData to retrieve
client sent Upgrade header. The virgin HttpRequest header is stored to
AccessLogEntry object.