Simplified by moving start conditions to startAutoConsumptionIfNeeded()
The original idea was to optimize by letting each caller to compute the
necessary conditions (because some of those conditions would not need to
be computed/checked in some callers). However, the end result looked too
inconsistent and increased the danger that some conditions may be
forgotten (or gotten wrong) in some callers.

Also removed double scheduleBodyDataNotification() calls (one explicit
and one via startAutoConsumption) in postAppend() as a side effect of
the above changes combined with further method code simplifications.