Revert commits 0040c029, 32c6d7ea  and bf39a4a5
This patches handle the case where squid retry a pconnrace on pinned
connection. In this case it is possible that the FwdState::startConnectionOrFail
method will try to operate on nil connection (which marks pinned connection)
causing segfaults to squid.
However this patch will cause squid to try reuse a closed pinned connection
resulting to a ZERO_SIZE_OBJECT/ServiceUnavailable  error to client side.

This is not what we are actually need:
 1) If repin is allowed then we should re-connect to the same server address
 2) Or at least propagate the server error to the client side

Currently there is not the mechanism for the (1) in squid and I  do not know
if we should try to build one into this issue.
The (2) is what the current squid trying to do, and probably what we
should still implement in this issue.