Fix check for pconnRace on pinned connections
Inside FwdState::supportsRepinning trying to detect pconn race on
Pinned connection checking if the FwdState::serverDestinations member has
a pinned item.
This is wrong because the FwdState::supportsRepinning maybe called after the
FwdState::serverDestications is modified to point to the next (and not pinned)
available path.

Replace with a simple check to FwdState::pconnRace.