Add a comment to explain stopAndDestroy call in FwdState::usePinned
The current squid policy about pinned connections, forces both client-side
and server-side connections to be closed on zero size replies.
Inside FwdState::usePinned the case the server-side connection found closed
just before squid start using it, is considered as a zero size reply case
and squid should just close client-side and server-side connections.

In practice, inside FwdState::usePinned, if the server-side pinned connection
found closed then the comm close handlers of pinned connections are already
scheduled for execution and maybe are already executed. In the case the
comm close handlers are already executed, the client-side connection
is not closed because the related zeroReply flag was not set (this is set
inside usePinned with the fail() call).
The stopAndDestroy call will trigger client-side connection to be closed too.