Bug 4919: master commit b599471 leaks memory
Restored the natural order of the following two notifications:
* BodyConsumer::noteMoreBodyDataAvailable() and
* BodyConsumer::noteBodyProductionEnded() or noteBodyProducerAborted().

Commit b599471 unintentionally reordered those two notifications. Client
kids (and possibly other BodyConsumers) relied on the natural order to
end their work. When noteBodyProductionEnded() came first, the "are we
done?" cleanup code in Client::handleAdaptedBodyProductionEnded() did
nothing, incorrectly assuming that the adapted body data has not been
consumed yet because we are still waiting for a resumeBodyStorage() call
(which has similar "are we done?" cleanup code). The other notification
handler, Client::handleMoreAdaptedBodyAvailable(), has no "are we done?"
cleanup code at all (because this method relies on a
noteBodyProductionEnded() follow up call to initiate cleanup).

If a HttpStateData job was done with the Squid-to-peer connection and
only waiting for the last adapted body bytes, it would get stuck and
leak many objects. This use case was not tested during b599471 work.