Polished new/renamed StatFlags documentation
I am not sure whether my definitions match the code because the code has
bugs (to be fixed later in this branch) and because two of the
definitions I am polishing (toProxy and peering) were the same. In the
subsequent commits, I will make the code match my definitions and/or
adjust the definitions.

I also did not like how easy it was to misinterpret "Request destined"
as the ultimate/final request destination rather than the next hop.

These polished definitions expose one or two problems:

1. We have two mutually exclusive boolean flags. That dangerous flag
   redundancy should be replaced by methods. (Marked as an XXX for now.)

2. The .peering flag seems to duplicate HttpStateData::_peer member.
   That duplication is necessary if non-HttpStateData users of
   StateFlags need this flag, but I have not checked whether they do.
   (Marked with a TODO for now.)