Bug 4928: Cannot convert non-IPv4 to IPv4. when reaching client_ip_max_connections
The client_ip_max_connections limit is checked before the TCP dst-IP is located
for the newely received TCP connection. This leaves Squid unable to fetch the
NFMARK or similar detaisl later on (they do not exist for [::]).

Move client_ip_max_connections test later in the TCP accept process to ensure
dst-IP is known when the error is produced.