Modify TunnelStateData code to use HappyConnOpener class
The HappyConnOpener::CbDialer class modified to be a template class.

Alternative implementation could be a HappyConnOpenerInitiator class:

class HappyConnOpenerInitiator: public virtual CbdataParent
{
 public:
    void noteConnection(const HappyConnOpener::Answer &) = 0;
};

Also the methods [FwdState/TunnelStateData]::{opening, notifyConnOpener} could
be merged to the common HappyConnOpenerInitiator class.
However these methods are not the only similar methods exists in FwdState
and TunnelStateData classes but also the PeerSelectionInitiator API is
very similar in these two classes.
Probably a base class for TunnelStateData and FwdStateData should be implemented
which will provide functionality for peer selection and connection opening.