Restored passing Connection by reference to GetMarkingsToServer()
For required arguments that are always dereferenced, we should favor
references instead of pointers, especially when the function does not
even check that the pointer is not nil.

Ip::Qos::setSockTos() and setSockNfmark() violate this principle (and
that violation probably caused removal of references in branch commit
b9f5414). I did not fix those function to avoid cascading changes.
However, I see no reason to propagate their interface problems to the
branch code we are responsible for.