Simplified HappyConnOpener::CbDialer, removing CbDialerBase
CbDialerBase should not have been derived from CallDialer because its
purpose is not provide the dialer API but to give us access to Answer.
It could be simplified to an AnswerCarrier helper class that has nothing
but an Answer data member. That was my first step.

Then I realized that if all our dialers are derived from Answer, then we
can cast our CallDialer pointer to Answer, eliminating the helper
AnswerCarrier class altogether. I believe this works even though Answer
itself is not polymorphic (as long as CallDialer remains polymorphic).