Fixed ErrorState leak and a design problem
Object owners should not be using CbcPointer for the objects they own
because CbcPointer does not own the object it points to. CbcPointer
counts references to the object, but it never initiates object deletion;
Object deletion inside CbcPointer may only happen as a side-effect of
decreasing the reference counter of an already invalidated (i.e.
essentially previously deleted) object.