Fixed a vector boundary overrun bug
The problem was that ProxyProtocol::PseudoFieldTypeToFieldName() tried
to access elements of an empty vector and used wrong index (raw enum
values) when accessing these elements.  Besides fixing this
problem I refactored a little, storing pseudo header names in a vector
instead of map. This is justified because the container has only few
elements and binary search gives no advantage in this case. Also this
approach does not require storing a static copy when performing reverse
search.