Fix builds using g++-4.8
Looks that the g++-4.8 defines the std::vector::erase to get non-const
iterator as parameter when compiles using c++11 language standard, instead
of a const iterator parameter defines the c++11 standard.

The extractFound method of ResolvedPeers class needs to get a non-const
iterator as parameter and the findSpareOrNextPeer should return a non-const
iterator too.
Unfortunatelly the fix required to convert the haveSpare and doneWithSpares
methods of ResolvedPeers class to be non-const.