Initialize preservingClientData_ at the start of the connection
... and possibly fix PROXY protocol support. We should never tunnel on
PROXY parsing failures because the PROXY prefix is required (where
expected) and is meant exactly for us, not some other peer.

Initializing when we set initial read timeout kind of makes sense -- if
we read nothing, we may indeed decide to tunnel that failure in case the
server wants to speak first, but it is not the setting of a read timeout
that should prompt a preservingClientData_ (re)set because setting the
timeout does not affect our ability to tunnel.

We should set preservingClientData_ when the connection is accepted
(because we may decide to start preserving bytes right away), and we
should reset it whenever that decision may change. This commit takes
care of the former. The existing code may already cover all the reset
cases, but confirming that is outside this commit scope.