Tried to address "request-count == 1" branch XXXs
The preserveHttpBytesForTunnellingUnsupportedProto() method is called
before parseOneRequest() ends, and, hence, before the request counters
are incremented in most cases. In those cases, counter 1 means that we
are parsing the "second" rather than the "first" request. This problem
may have been introduced by my recent branch commits -- I have not
checked when it appeared first.

I know of only one case where the plain request counter is incremented
before ConnStateData::parseOneRequest(): buildFakeRequest(). AFAICT,
buildFakeRequest*( code paths always result in tunneling. If we made the
decision to tunnel, then the client data should have been preserved
already. Thus, preserveHttpBytesForTunnellingUnsupportedProto() is
probably not even called after the request counter is incremented in
such cases.

Also placed the counter check first so that the order of checks matches
the description of the if-statement condition: "the first request on a
... connection".