A better preserveHttpBytesForTunnellingUnsupportedProto() name
IIRC, at some point, the method was dealing with supposed-to-be-HTTP
bytes exclusively, but it is now also handling supposed-to-be-TLS bytes.
Several other methods and data members use "client data" terminology.

Using Client Bytes would be better, but that would require renaming
preservedClientData which we are not otherwise changing on this branch.