RFC 8586: Loop Detection in Content Delivery Networks (#396)
Support the CDN-Loop header as a source for loop detection.

This header is only relevant to CDN installations. For which the
surrogate_id configuration directive specifies the authoritative
ID.

Squid does not add this header by default, preferring to use the
Via mechanism instead. Administrators may add it to requests
with the request_header_add directive or remove with
request_header_remove.