Fix "BUG: Lost previously bumped from-Squid connection..."
While forwarding bumped requests the peer selection procedure may either
found the pinned connection closed (or closing) or the peer selection
acls prevent from using the cache peer used by pinned connection.
The above cases are not refer to real bugs as the debug messages says and
just close the connection (as the existing squid code does) is the correct
behavior.

This patch:
  - does not check for to-server pinned connection health inside
    PeerSelector. The PeerSelectionInitiator is responsible to check
    for pinned connections health.

  - Add code to FwdState class to check if the peer selection procedure
    does not allow to use existing pined connection but require a new
    connection. In this case check if the type of pinned connection
    allow retries and if yes unpin the existing to-server connection and
    start a new one to the server.

This is a Measurement Factory project