Remove URN force menu hack
The hack was that clients could force Squid to produce a menu
page listing URN alternative URL mapings without a Location
header auto-redirecting to the fastest one.

A typo was added in July 2015 (Squid-3.5.7) which broke the
hack by altering the trigger string. Nobody noticed.

Given the '.' used by this hack is not a valid NID character
it seems a good time to drop this hack entirely rather than
twist the new parser to cope with it. If it turns out we do
need this behaviour a new hack will have to be designed that
complies with URN syntax.