Remove userinfo support from old protocols
RFC 1738 defines the URL schemes for gopher and wais as not
having the userinfo@ segment.

whois: and cache_object: schemes also do not use this segment.

For these cases we can obey the RFC7230 requirement to ignore the
segment when producing normalized absolute URI.

Of the supported protocols only FTP requires userinfo, and because
we cannot be certain for unknown protocols allow it as well.