Check only base upgrade protocols on server response evaluation.
This is allow the following squid upgrade requests and server responses pairs:
 1) request upgrade to protocol/X and server responses with protocol/Y
 2) request upgrade to protocol and server responses with protocol/Y
 3) request upgrade to protocol/X and server responses with protocol

While squid evaluates the http_upgrade_request_protocols acls still takes
in account the protocol releases. This is allow admin to have a control
over the advertised upgrade protocols.
In the case the admin wants to deny upgrading to special releases of a
protocol he should (be able to) block the related HTTP/101 responses.