Update FwdState::serverConn only after the connection can be used
It is wrong to have a FwdState::serverConn without close handler.  Keep
this member nil untill the Http::Tunneler and Ssl::PeerConnector objects
are finished and the connection can be realy used.
The same adjustions required for the tunnel.cc code to.

This patch also adds the Http::Tunneler::usesPconn_ and
Security::PeerConnector::usesPconn_ flags to mark that their objects
should update the fwdPconnPool. These flags are similar with the
HappyConnOpener::mayUsesPconn_ flag added recently to squid master.