Replaced ignorant const-violating getReply() code
... by adding adjustableBaseReply() accessor methods. The alternative
is to require replaceBaseReply() calls (with freshly created replies).
When applied to incremental parsing, that alternative would require the
caller to maintain its own HttpReply structure. However, it should still
be seriously considered because it is architecturaly simpler and could,
with additional refactoring, remove the overhead of creating numerous
empty replies that are overwritten by replaceBaseReply() calls (TODO).