Mark callers that need stable reply properties with stableReply()
This approach works well for callers that do not care much about which
reply to use (base or freshest), as intended.

However, it also creates a conflict for callers that focus on receiving
data from Core and, hence, should be using baseReply(). The same caller
may also access presumably stable reply properties so they are tempted
to use stableReply(), creating a (mild) call selection conflict.