Throw Esi::ErrorDetail instead of generic TextException
This allows us to better distinguish between detected
issues handling the parser state vs other system issues.
The former can be delivered externally to clients the
latter should only be logged for the admin.

SBuf is used for now as the only detail needing to be thrown
is the error message text and the code throwing does not
have access to the HttpRequest and AccessLogEntry objects
needed to create an ErrorState.